reduce size a bit

+++ test/distance/examples/osa.hpp
@@ -1,5 +1,5 @@
 #pragma once
 #include <string>
 
-extern std::basic_string<uint32_t> osa_example1;
-extern std::basic_string<uint32_t> osa_example2;
+extern std::basic_string<uint8_t> osa_example1;
+extern std::basic_string<uint8_t> osa_example2;
